+++ src/rust/src/utils/mod.rs
@@ -1,4 +1,3 @@
 pub mod general;
 pub mod macros;
 pub mod r_rust_interface;
-pub mod utils_stats;