+++ docs/change_log.md
@@ -2,6 +2,13 @@
 
 *Last update: 04.09.2025* </br>
 
+### Version **0.0.2.2**
+
+#### New features
+
+- Mitch multi contrast enrichment in Rust.
+- Rework and -factor of sub modules in the Rust code.
+
 ### Version **0.0.2.1**
 
 #### New features